+++ docs/docs_app/custom.py
@@ -1,6 +1,5 @@
 import dash_core_components as dcc
 import dash_html_components as html
-
 import dash_loading_spinners as dls
 
 svg = """
