+++ docs/docs_app/examples.py
@@ -5,10 +5,10 @@
 import dash_bootstrap_components as dbc
 import dash_core_components as dcc
 import dash_html_components as html
-from dash.dependencies import Input, Output
-from helpers import app
-
 import dash_loading_spinners as dls
+from dash.dependencies import Input, Output
+
+from docs_app.app import app
 
 spinner_options = {
     "Dots": {
