+++ docs/setup.cfg
@@ -1,0 +1,16 @@
+[metadata]
+name = docs_app
+description = Documentation for dash-loading-spinners
+license = MIT
+version = 0.1.0
+author = glsdown
+author_email = gemmalsdown@icloud.com
+url = https://github.com/glsdown/dash-loading-spinners
+
+[options]
+packages=find:
+python_requires = >=3.9
+install_requires =
+    dash
+    dash-bootstrap-components
+    numpy
